+++ Accounts/supplychain/gradle.properties
@@ -1,3 +1,3 @@
-name=Test
+name=Supplychain Cordapp
 group=com.supplychain
-version=0.1
+version=1.0
