+++ Accounts/supplychain/workflows/src/main/java/net/corda/samples/supplychain/flows/SendCargo.java
@@ -1,0 +1,114 @@
+package net.corda.samples.supplychain.flows;
+
+import co.paralleluniverse.fibers.Suspendable;
+import com.r3.corda.lib.accounts.contracts.states.AccountInfo;
+import com.r3.corda.lib.accounts.workflows.services.AccountService;
+import com.r3.corda.lib.accounts.workflows.flows.RequestKeyForAccount;
+import com.r3.corda.lib.accounts.workflows.services.KeyManagementBackedAccountService;
+import com.sun.istack.NotNull;
+import net.corda.samples.supplychain.accountUtilities.NewKeyForAccount;
+import net.corda.samples.supplychain.contracts.CargoStateContract;
+import net.corda.samples.supplychain.states.CargoState;
+import net.corda.core.crypto.TransactionSignature;
+import net.corda.core.flows.*;
+import net.corda.core.identity.AnonymousParty;
+import net.corda.core.identity.Party;
+import net.corda.core.transactions.SignedTransaction;
+import net.corda.core.transactions.TransactionBuilder;
+
+import java.security.PublicKey;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+
+// ******************
+// * Initiator flow *
+// ******************
+@InitiatingFlow
+@StartableByRPC
+public class SendCargo extends FlowLogic<String> {
+
+    //private variables
+    private String pickupFrom ;
+    private String whereTo;
+    private String cargo;
+
+
+    //public constructor
+    public SendCargo(String pickupFrom, String shipTo, String cargo){
+        this.pickupFrom = pickupFrom;
+        this.whereTo = shipTo;
+        this.cargo = cargo;
+    }
+
+    @Suspendable
+    @Override
+    public String call() throws FlowException {
+        //grab account service
+        AccountService accountService = getServiceHub().cordaService(KeyManagementBackedAccountService.class);
+        //grab the account information
+        AccountInfo myAccount = accountService.accountInfo(pickupFrom).get(0).getState().getData();
+        PublicKey myKey = subFlow(new NewKeyForAccount(myAccount.getIdentifier().getId())).getOwningKey();
+
+        AccountInfo targetAccount = accountService.accountInfo(whereTo).get(0).getState().getData();
+        AnonymousParty targetAcctAnonymousParty = subFlow(new RequestKeyForAccount(targetAccount));
+
+        //generating State for transfer
+        CargoState output = new CargoState(new AnonymousParty(myKey),targetAcctAnonymousParty,cargo,getOurIdentity());
+
+        // Obtain a reference to a notary we wish to use.
+        /** METHOD 1: Take first notary on network, WARNING: use for test, non-prod environments, and single-notary networks only!*
+         *  METHOD 2: Explicit selection of notary by CordaX500Name - argument can by coded in flow or parsed from config (Preferred)
+         *
+         *  * - For production you always want to use Method 2 as it guarantees the expected notary is returned.
+         */
+        final Party notary = getServiceHub().getNetworkMapCache().getNotaryIdentities().get(0); // METHOD 1
+        // final Party notary = getServiceHub().getNetworkMapCache().getNotary(CordaX500Name.parse("O=Notary,L=London,C=GB")); // METHOD 2
+
+        TransactionBuilder txbuilder = new TransactionBuilder(notary)
+                .addOutputState(output)
+                .addCommand(new CargoStateContract.Commands.Create(), Arrays.asList(targetAcctAnonymousParty.getOwningKey(),getOurIdentity().getOwningKey()));
+
+        //self sign Transaction
+        SignedTransaction locallySignedTx = getServiceHub().signInitialTransaction(txbuilder,Arrays.asList(getOurIdentity().getOwningKey()));
+
+        //Collect sigs
+        FlowSession sessionForAccountToSendTo = initiateFlow(targetAccount.getHost());
+        List<TransactionSignature> accountToMoveToSignature = (List<TransactionSignature>) subFlow(new CollectSignatureFlow(locallySignedTx,
+                sessionForAccountToSendTo,targetAcctAnonymousParty.getOwningKey()));
+        SignedTransaction signedByCounterParty = locallySignedTx.withAdditionalSignatures(accountToMoveToSignature);
+
+        //Finalize
+        subFlow(new FinalityFlow(signedByCounterParty,
+                Arrays.asList(sessionForAccountToSendTo).stream().filter(it -> it.getCounterparty() != getOurIdentity()).collect(Collectors.toList())));
+        return "send " + cargo+ " to " + targetAccount.getHost().getName().getOrganisation() + "'s "+ targetAccount.getName() + " team";
+
+    }
+}
+
+
+@InitiatedBy(SendCargo.class)
+class SendCargoResponder extends FlowLogic<Void> {
+    //private variable
+    private FlowSession counterpartySession;
+
+    //Constructor
+    public SendCargoResponder(FlowSession counterpartySession) {
+        this.counterpartySession = counterpartySession;
+    }
+
+    @Override
+    @Suspendable
+    public Void call() throws FlowException {
+        subFlow(new SignTransactionFlow(counterpartySession) {
+            @Override
+            protected void checkTransaction(@NotNull SignedTransaction stx) throws FlowException {
+                // Custom Logic to validate transaction.
+            }
+        });
+        subFlow(new ReceiveFinalityFlow(counterpartySession));
+        return null;
+    }
+}
+
