--- Accounts/supplychain/workflows/src/test/java/com/supplychain/ContractTests.java
@@ -1,13 +1,0 @@
-package com.supplychain;
-
-import net.corda.testing.node.MockServices;
-import org.junit.Test;
-
-public class ContractTests {
-    private final MockServices ledgerServices = new MockServices();
-
-    @Test
-    public void dummyTest() {
-
-    }
-}
